spell: increase move point limit
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2292 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/spell/spell.cpp
@@ -38,7 +38,7 @@ struct spellstats_t
 {
         const char* name;
         u8 sp;
-        u8 mp;
+        u16 mp;
         u8 sprite;
         u8 extra;
 	u8 bits;
